refactor: restructured some folders
