+++ src/commands/fun/StufiCommand.ts
