+++ src/commands/meme/DeepFryCommand.ts
