+++ src/commands/meme/SpooktoberCommand.ts
