+++ src/commands/text/MockifyCommand.ts
