+++ src/index.ts
@@ -8,20 +8,20 @@ import dayjs from 'dayjs';
 import utc from 'dayjs/plugin/utc';
 import timezone from 'dayjs/plugin/timezone';
 import {Maybe} from '@/types/util';
-import {KortebroekCommand} from '@/commands/KortebroekCommand';
+import {KortebroekCommand} from '@/commands/fun/KortebroekCommand';
 import {PingCommand} from '@/commands/PingCommand';
-import {StufiCommand} from '@/commands/StufiCommand';
+import {StufiCommand} from '@/commands/fun/StufiCommand';
 import {HelpCommand} from '@/commands/HelpCommand';
-import {WhereMemeCommand} from '@/commands/WhereMemeCommand';
-import {IAmDadEvent} from './events/IAmDadEvent';
-import {BangerEvent} from './events/BangerEvent';
-import {SpooktoberCommand} from './commands/SpooktoberCommand';
-import {DeepFryCommand} from './commands/DeepFryCommand';
-import {JokeMemeCommand} from './commands/JokeMemeCommand';
-import {MockifyCommand} from './commands/MockifyCommand';
-import {DrakeMemeCommand} from './commands/DrakeMemeCommand';
-import {UwuCommand} from './commands/UwuCommand';
-import {MarieKondoCommand} from './commands/MarieKondoCommand';
+import {WhereMemeCommand} from '@/commands/meme/WhereMemeCommand';
+import {IAmDadEvent} from './events/meme/IAmDadEvent';
+import {BangerEvent} from './events/meme/BangerEvent';
+import {SpooktoberCommand} from './commands/meme/SpooktoberCommand';
+import {DeepFryCommand} from './commands/meme/DeepFryCommand';
+import {JokeMemeCommand} from './commands/meme/JokeMemeCommand';
+import {MockifyCommand} from './commands/text/MockifyCommand';
+import {DrakeMemeCommand} from './commands/meme/DrakeMemeCommand';
+import {UwuCommand} from './commands/text/UwuCommand';
+import {MarieKondoCommand} from './commands/meme/MarieKondoCommand';
 import {I18n} from 'i18n';
 
 const DISCORD_APPLICATION_ID = process.env.DISCORD_APPLICATION_ID ?? '';
